refactor: pages 폴더 및 불필요한 타입 파일 삭제
