+++ src/_components/Setting/index.tsx
@@ -5,7 +5,7 @@ import Button from '@/_components/common/Button/Button';
 import Nav from '@/_components/common/Nav/Nav';
 import PageInnerLayout from '@/_components/common/PageInnerLayout/PageInnerLayout';
 import { MAIL_FREQUENCY } from '@/_constants/setting';
-import { MailFrequency } from '@/_types/setting';
+import { MailFrequency } from '@/_types';
 import * as React from 'react';
 import MailFrequencySection from './MailFrequencySection';
 import { settingButton, settingContainer, settingContent, settingTitle } from './setting.css';
