--- src/_types/setting.ts
@@ -1,3 +1,0 @@
-import { MAIL_FREQUENCY } from './../_constants/setting';
-
-export type MailFrequency = keyof typeof MAIL_FREQUENCY;
